+++ website_local_content/tests/__init__.py
@@ -1,5 +1,6 @@
 from . import (
     test_local_content_controller,
     test_local_content_models,
+    test_local_content_rating,
     test_local_content_tag,
 )